fix: use BigDecimal instead of double

+++ entities/src/main/java/com/spark/entities/domain/ProductDTO.java
@@ -3,12 +3,11 @@ package com.spark.entities.domain;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.NotEmpty;
 
-import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -21,5 +20,5 @@ public class ProductDTO {
     private String description;
     @NotNull(message = "Price is required")
     @Min(value = 0, message = "Price must be greater than or equal to zero")
-    private double price;
+    private BigDecimal price;
 }
